requirements.txt
accelerate==0.27.0
cffi
decord==0.6.0
deepspeed==0.14.4
editdistance
jmespath
Levenshtein
numpy==1.26.4
openai
Pillow==10.1.0
protobuf==4.25.0
pytrec_eval==0.5
sentencepiece==0.1.99
torch==2.1.2
torchvision==0.16.2
transformers==4.40.2
